CHANGE: Default to only adding site-packages of the env.
I'm thinking that it might still be necessary to add the full stdlib
sys.path for an env, because e.g. there are API changes in py3 as
compared to py2. So if the calling syntax for the stdlib stuff for
py3 is different, it would be nice if jedi-vim presents the py3 stuff
for completion.  Dunno, we'll have to see how that goes.